lint
SCCS-vsn: sys/kern/kern_physio.c 4.26
SCCS-vsn: sys/kern/vfs_bio.c 4.26
SCCS-vsn: sys/kern/vfs_cluster.c 4.26
SCCS-vsn: sys/vax/vax/cons.c 4.13
SCCS-vsn: sys/vax/uba/dz.c 4.35
SCCS-vsn: sys/kern/tty_pty.c 4.19
SCCS-vsn: sys/vax/if/if_acc.c 4.9
SCCS-vsn: sys/vax/if/if_dmc.c 4.4
SCCS-vsn: sys/vax/if/if_en.c 4.38
SCCS-vsn: sys/deprecated/netimp/if_imp.c 4.14
SCCS-vsn: sys/netinet/in_pcb.c 4.20
SCCS-vsn: sys/deprecated/netpup/pup.h 4.3
SCCS-vsn: sys/net/raw_cb.c 4.5
SCCS-vsn: sys/deprecated/netimp/raw_imp.c 4.8
SCCS-vsn: sys/netinet/raw_ip.c 4.7
SCCS-vsn: sys/deprecated/netpup/raw_pup.c 4.8
SCCS-vsn: sys/net/raw_usrreq.c 4.9
SCCS-vsn: sys/netinet/tcp_debug.c 4.2
SCCS-vsn: sys/netinet/tcp_input.c 1.60
SCCS-vsn: sys/netinet/tcp_output.c 4.33
SCCS-vsn: sys/netinet/tcp_subr.c 4.17
SCCS-vsn: sys/kern/uipc_mbuf.c 1.32
SCCS-vsn: sys/kern/vfs_lookup.c 4.13
SCCS-vsn: sys/ufs/ffs/ufs_lookup.c 4.13
SCCS-vsn: sys/ufs/ufs/ufs_lookup.c 4.13
SCCS-vsn: sys/kern/uipc_pipe.c 4.11
SCCS-vsn: sys/kern/uipc_proto.c 4.18

+++ usr/src/sys/net/raw_usrreq.c
@@ -1,4 +1,4 @@
-/*	raw_usrreq.c	4.8	82/03/05	*/
+/*	raw_usrreq.c	4.9	82/03/13	*/
 
 #include "../h/param.h"
 #include "../h/mbuf.h"
@@ -128,7 +128,7 @@ next:
 		if (last) {
 			struct mbuf *n;
 
-			if (n = m_copy(m->m_next, 0, M_COPYALL))
+			if (n = m_copy(m->m_next, 0, (int)M_COPYALL))
 				goto nospace;
 			if (sbappendaddr(&last->so_rcv, &rawp->raw_src, n) == 0) {
 				/*
